feat(sdl): move updateEmail to new pattern as example
